coding/game-art-visualizer/requirements.txt
streamlit
requests
python-dotenv
pillow
colorthief
matplotlib
pandas
numpy
scikit-learn
